+++ CHANGELOG.md
@@ -39,6 +39,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - Deprecate `@cumulus/common/concurrency.promiseUrl`
   - Deprecate `@cumulus/common/concurrency.toPromise`
   - Deprecate `@cumulus/common/concurrency.unless`
+  - Deprecate `@cumulus/common/config.buildSchema`
+  - Deprecate `@cumulus/common/config.parseConfig`
+  - Deprecate `@cumulus/common/config.resolveResource`
+  - Deprecate `@cumulus/common/config.resourceToArn`
   - Deprecate `@cumulus/common/Semaphore`
 
 ### Removed